msr: Adding check for kernel module
Previously, if the kernel module was not present, this was silently
returning an empty file-handle. When the MSR kernel module is not
present, we want to make sure a specific error is returned letting the
user know that it is not present. For now, we are checking this by
looking if the device-file is present.

Signed-off-by: Larry Dewey <larry.dewey@amd.com>